Add card flipping on click
Cards now flip over when clicked. However, the counter seems to trigger
the useEffect in charge of shuffling the cards even if it is set
to update after the cards have been flipped. This results in a graphical
glitch where users can see new card positions before the flip effect
completes.